fixes failing tests

+++ RUFAS/biophysical/animal/milk/milk_production.py
@@ -262,7 +262,7 @@ class MilkProduction:
         [AN.MLK.9]
 
         """
-        return float(l_param * np.power(days_in_milk, m_param) * np.exp(-1 * n_param * days_in_milk))
+        return l_param * np.power(days_in_milk, m_param) * np.exp(-1 * n_param * days_in_milk)
 
     @staticmethod
     def calculate_predicted_305_day_milk_yield(l_param: float, m_param: float, n_param: float) -> float:
@@ -292,7 +292,7 @@ class MilkProduction:
 
         """
         result, _ = quad(MilkProduction.calculate_daily_milk_production, 1, 305, args=(l_param, m_param, n_param))
-        return float(result)
+        return result
 
     def calculate_305_day_milk_yield(self) -> float:
         """
